[GEOS-11249] Split COG community module packaging based on target cloud provider
