+++ src/community/cog/cog-azure/src/test/java/org/geoserver/rest/catalog/CogAzureRemoteHarvestOnlineTest.java
