+++ src/community/cog/cog-core/src/main/java/applicationContext.xml
