+++ src/community/cog/cog-core/src/main/java/org/geoserver/cog/CogSettingsStore.java
