+++ src/community/cog/cog-core/src/main/java/org/geoserver/web/data/store/cog/panel/CogRasterEditPanel.java
