+++ src/community/cog/cog-core/src/main/resources/GeoServerApplication_fr.properties
