+++ src/community/cog/cog-google/src/assembly/assembly.xml
@@ -1,0 +1,31 @@
+<assembly>
+  <id>cog-google-plugin</id>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <includeBaseDirectory>false</includeBaseDirectory>
+  <fileSets>
+    <fileSet>
+      <directory>target/dependency</directory>
+      <outputDirectory></outputDirectory>
+      <includes>
+        <include>gs-cog*.jar</include>
+        <include>imageio-ext-*rangereader*.jar</include>
+        <include>*google*.jar</include>
+        <include>guava*.jar</include>
+        <include>checker-qual*.jar</include>
+        <include>error_prone_annotations*.jar</include>
+        <include>objc-annotations*.jar</include>
+        <include>gson*.jar</include>
+        <include>auto-value-annotations*</include>
+        <include>gax*</include>
+        <include>api-common*</include>
+        <include>annotation-api*</include>
+        <include>grpc-context*</include>
+        <include>opencensus*</include>
+        <include>protobuf-java*</include>
+        <include>threetenbp*</include>
+      </includes>
+    </fileSet>
+  </fileSets>
+</assembly>
