+++ src/community/cog/cog-google/src/test/java/org/geoserver/rest/catalog/CogGSRemoteHarvestOnlineTest.java
