+++ src/community/cog/cog-s3/src/assembly/assembly.xml
@@ -1,0 +1,51 @@
+<assembly>
+  <id>cog-s3-plugin</id>
+  <formats>
+    <format>zip</format>
+  </formats>
+  <includeBaseDirectory>false</includeBaseDirectory>
+  <fileSets>
+    <fileSet>
+      <directory>target</directory>
+      <outputDirectory></outputDirectory>
+      <includes>
+        <include>${artifactId}-${project.version}.jar</include>
+      </includes>
+    </fileSet>
+    <fileSet>
+      <directory>target/dependency</directory>
+      <outputDirectory></outputDirectory>
+      <includes>
+        <include>annotations-2.*.jar</include>
+        <include>auth-2.*.jar</include>
+        <include>aws-*-2.*.jar</include>
+        <include>azure-storage*.jar</include>
+        <include>client-runtime*.jar</include>
+        <include>gs-cog*.jar</include>
+        <include>http-client-spi-2*.jar</include>
+        <include>http*-osgi*.jar</include>
+        <include>imageio-ext-*rangereader*.jar</include>
+        <include>jackson-datatype*.jar</include>
+        <include>jackson-module*.jar</include>
+        <include>kotlin-stdlib*.jar</include>
+        <include>*netty*.jar</include>
+        <include>okhttp*.jar</include>
+        <include>okio*.jar</include>
+        <include>profiles-2.*.jar</include>
+        <include>protocol-core-2.*.jar</include>
+        <include>reactive-streams*.jar</include>
+        <include>regions-2.*.jar</include>
+        <include>rxjava*.jar</include>
+        <include>s3-2.*.jar</include>
+        <include>sdk-core-2.*.jar</include>
+        <include>utils-2.*.jar</include>
+        <include>checksums-*2*.jar</include>
+        <include>crt-core-*-2*.jar</include>
+        <include>endpoints-spi-2*.jar</include>
+        <include>http-auth-*2*.jar</include>
+        <include>identity-spi*.jar</include>
+        <include>metrics-spi-2.*.jar</include>
+      </includes>
+    </fileSet>
+  </fileSets>
+</assembly>
