+++ src/community/cog/cog-s3/src/test/java/org/geoserver/rest/catalog/CogRemoteHarvestOnlineTest.java
